Removed unwanted files. Refactored components
