+++ src/index.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import './index.css';
-import App from './App';
-import { store } from './app/store';
+import 'styles/index.css';
+import App from 'components/App';
+import { store } from 'store/store';
 import { Provider } from 'react-redux';
 import * as serviceWorker from './serviceWorker';
